core-2001 allow recipe

+++ src/Spryker/Deploy/Timer/Timer.php
@@ -35,6 +35,6 @@ class Timer implements TimerInterface
     {
         $start = $this->timer[spl_object_hash($object)];
 
-        return sprintf('%.2f', microtime(true) - $start);
+        return (float)sprintf('%.2f', microtime(true) - $start);
     }
 }
